+++ src/app/routes/App/components/MonitorStatus.tsx
@@ -18,45 +18,26 @@
 
 // Vendor
 import * as React from 'react';
-import { bindActionCreators } from 'redux';
-import { connect } from 'react-redux';
 
 // Local
-import { MonitorModal } from './MonitorModal';
-import { NormalizedMonitorList } from '../../../services/monitors/types';
-import {
-  StateToProps,
-  DispatchToProps,
-} from '../../../types/redux';
-import {
-  fetchMonitors,
-  selectMonitor,
-  closeModal,
-  openModal,
-} from '../actions/MonitorStatusActions';
+import { MonitorModalContainer } from '../containers/MonitorModalContainer';
 
 // --------------------------------------------------------------------------
 // Interfaces/Types
 // --------------------------------------------------------------------------
 
 /** Value properties of the MonitorStatus component. */
-interface ValueProps {
-  /** If a loading icon should be shown in the monitor modal. */
-  loading: boolean;
-  /** Indexed list of current monitor objects. */
-  monitors: NormalizedMonitorList | null;
+export interface ValueProps {
   /** Monitors that are currently running. */
   monitorsUp: string[];
   /** Monitors that are currently down. */
   monitorsDown: string[];
-  /** If the monitor modal is active. */
-  modalActive: boolean;
-  /** Selected monitor from the current list of indexed monitors. */
-  selectedMonitor: string | null;
+  /** Timeout ID of the monitor polling. */
+  pollTimeoutID: number | null;
 }
 
 /** Function properties of the MonitorStatus component. */
-interface FunctionProps {
+export interface FunctionProps {
   /**
    * Selects a monitor from the list.
    * @param monitor ID of the monitor to select.
@@ -67,7 +48,7 @@ interface FunctionProps {
   /** Opens the monitor modal. */
   openModal(): any;
   /** Gets the current list of monitors. */
-  fetchMonitors(): any;
+  fetchMonitors(loading: boolean, delay: number, timeoutID?: number): any;
 }
 
 /** Combined property interfaces for the MonitorStatus component. */
@@ -82,67 +63,43 @@ type Props = ValueProps & FunctionProps;
  * clicked displays a modal with the information of all the monitors.
  */
 export class MonitorStatus extends React.Component<Props, {}> {
-  /**
-   * Fetches the current monitor list.
-   */
+  /** Time to wait before polling for new monitors in milliseconds. */
+  public static POLLING_DELAY = 60000;
+
   public componentWillMount(): void {
-    this.props.fetchMonitors();
+    this.fetchMonitors(false);
   }
+
+  /**
+   * Fetches the current list of monitors and starts the monitor poll.
+   * @param loading If a loading icon should be shown on the monitor modal.
+   * @param timeoutID ID of the polling timeout.
+   */
+  public fetchMonitors = (loading: boolean, timeoutID?: number): void => {
+    this.props.fetchMonitors(
+      loading,
+      MonitorStatus.POLLING_DELAY,
+      timeoutID,
+    );
+  };
+
+  /** Fetches the monitors and opens the modal. */
+  public openModal = (): void => {
+    this.fetchMonitors(true, this.props.pollTimeoutID || undefined);
+    this.props.openModal();
+  };
 
   public render(): JSX.Element {
     return (
       <div className="flex-item flex--shrink">
-        <a className="header__link" onClick={this.props.openModal}>
+        <a className="header__link" onClick={this.openModal}>
           {this.props.monitorsUp.length + ' '} <i className="fa fa-arrow-up" />
           {' '}
           {this.props.monitorsDown.length + ' '} <i className="fa fa-arrow-down" />
 
-          <MonitorModal
-            loading={this.props.loading}
-            closeModal={this.props.closeModal}
-            modalActive={this.props.modalActive}
-            monitors={this.props.monitors}
-            monitorsDown={this.props.monitorsDown}
-            monitorsUp={this.props.monitorsUp}
-            selectMonitor={this.props.selectMonitor}
-            selectedMonitor={this.props.selectedMonitor}
-          />
+          <MonitorModalContainer />
         </a>
       </div>
     );
   }
 }
-
-// --------------------------------------------------------------------------
-// Container
-// --------------------------------------------------------------------------
-
-/**
- * Maps the current redux state to the MonitorStatus component.
- * @param state
- */
-const values: StateToProps<ValueProps, undefined> = (state) => ({
-  loading: state.routes.App.MonitorStatus.loading,
-  modalActive: state.routes.App.MonitorStatus.modalActive,
-  monitors: state.routes.App.MonitorStatus.monitors,
-  monitorsDown: state.routes.App.MonitorStatus.monitorsDown,
-  monitorsUp: state.routes.App.MonitorStatus.monitorsUp,
-  selectedMonitor: state.routes.App.MonitorStatus.selectedMonitor,
-});
-
-/**
- * Maps redux dispatch actions to the MonitorStatus component.
- * @param dispatch
- */
-const functions: DispatchToProps<FunctionProps, undefined> = (dispatch) => ({
-  closeModal: bindActionCreators(closeModal, dispatch),
-  fetchMonitors: bindActionCreators(fetchMonitors, dispatch),
-  openModal: bindActionCreators(openModal, dispatch),
-  selectMonitor: bindActionCreators(selectMonitor, dispatch),
-});
-
-/** Container of the MonitorStatus component. */
-export const MonitorStatusContainer = connect(
-  values,
-  functions,
-)(MonitorStatus);